General code re-organisation
This commit does a lot (sorry about that).

- I rewrote the i/o from files functions they now lives in src/io.c.
- I moved the node movements related functions in their own src/movements.c
file to reduce the size of src/partition.c.
- I fixed the normalisation and filling of AdjaArray/Partitions and put
everything in src/fillpartitions.c
- I removed some old code that used to live en src/costmatrix.c (to convert
from the legacy datastructure, now that we have the new i/o it was
useless) and src/louvain.c (it was broken anyway).
