+++ src/Makefile.am
@@ -1,5 +1,5 @@
-INCLUDES		=  -I$(top_builddir) -I$(top_srcdir) 
+INCLUDES		=  -I$(top_builddir) -I$(top_srcdir)
 
 pkglib_LIBRARIES 	= librgraph.a
-librgraph_a_SOURCES = datastruct.c tools.c graph.c models.c modules.c bipartite.c missing.c matrix.c layout.c recommend.c conflict.c genetic.c louvain.c sannealing.c costmatrix.c partition.c io.c
-pkginclude_HEADERS 	= datastruct.h tools.h graph.h models.h modules.h bipartite.h missing.h matrix.h layout.h recommend.h conflict.h genetic.h louvain.h sannealing.h costmatrix.h partition.h io.h
+librgraph_a_SOURCES = datastruct.c tools.c graph.c models.c modules.c bipartite.c missing.c matrix.c layout.c recommend.c conflict.c genetic.c sannealing.c partition.c io.c fillpartitions.c io.c movements.c
+pkginclude_HEADERS 	= datastruct.h tools.h graph.h models.h modules.h bipartite.h missing.h matrix.h layout.h recommend.h conflict.h genetic.h sannealing.h partition.h fillpartitions.h io.h movements.h
