+++ src/io.c
@@ -1,41 +1,214 @@
-#include <stdlib.h>
+#include <stdio.h>
 #include <math.h>
+#include <stdlib.h>
 #include <search.h>
 #include <string.h>
-#include <gsl/gsl_rng.h>
 
+#include <gsl/gsl_rng.h>
 #include "partition.h"
 #include "io.h"
 
+int
+node_compare(const void *node1, const void *node2)
+{
+    return strcmp(((const struct node *) node1)->label,
+        ((const struct node *) node2)->label);
+}
+
+
+struct node_t
+{
+  /* Callers expect this to be the first element in the structure - do not
+     move!  */
+  const void *key;
+  struct node_t *left;
+  struct node_t *right;
+  unsigned int red:1;
+};
+/**
+@brief Free the memory allocated to binary tree.
+*/
+void
+free_node_tree(void *root)
+{
+  struct node_t *focal = (struct node_t*) root;
+  if (focal != NULL){
+	struct node *leaf = (struct node*) focal->key;
+   	free_node_tree(focal->left);
+   	free_node_tree(focal->right);
+	free(leaf);
+	free(focal);
+  }
+}
+
 
 /**
-  Returns the role number given a (P,z) tuple.
-
-  Cf. Guimera & Amaral, Nature (2005) for the reasoning behind the
-  boundaries values.
+Read an edge list from a file and convert it to arrays.
+@param nodes_in_p,nodes_out_p,weights_p Edge list, nodes id and weight of each edge.
+@param labels_p Labels of nodes whose index are in nodes_in/out
+@param E_p N_p Number of edges, number of nodes.
+@param weighted If True
+@param bipartite 0,1,2. If 1 (resp 2), treat the conlumns independently and store only the first (resp 2nd) column labels.
 **/
-int GetRole(double P, double z){
-  int dest_group;
-  if (z < 2.5) {  /* Node is not a hub */
-	if (P < 0.050)
-	  dest_group = 0;
-	else if (P < 0.620)
-	  dest_group = 1;
-	else if (P < 0.800)
-	  dest_group = 2;
-	else
-	  dest_group = 3;
+int
+EdgeListFileInput(FILE *inFile, int weighted, int bipartite, unsigned int **nodes_in_p,
+                  unsigned int **nodes_out_p,double **weights_p,
+							    char ***labels_p, unsigned int *E_p, unsigned int *N_p){
+
+  unsigned int *nodes_in=NULL, *nodes_out=NULL;
+  double *weights=NULL;
+  char **labels=NULL;
+
+  if (bipartite > 2) return 2;
+
+  int i = 0;
+
+  int noReadItems;
+  char *line = NULL;
+  size_t bufsiz = 0;
+  ssize_t nbytes;
+
+  unsigned int Nmax = 10;
+  unsigned int Emax = Nmax*Nmax;
+
+  labels = calloc(Nmax+1,sizeof(char*));
+  weights = malloc((Emax+1)*sizeof(double));
+  nodes_in = malloc((Emax+1)*sizeof(unsigned int));
+  nodes_out = malloc((Emax+1)*sizeof(unsigned int));
+
+  unsigned int N = 0;
+  unsigned int N_other_side = 0;
+  unsigned int E = 0;
+
+  struct node *nodes_root = NULL;
+  struct node *nodes_root_bipartite = NULL;
+
+  struct node *nodeptr = NULL;
+
+  unsigned int noReadItemsExpected = weighted ? 3:2;
+
+  // Go through the input file
+  while ((nbytes = getline(&line, &bufsiz, inFile)) != -1){
+    //printf("N=%d/%d, E=%d/%d\n",N,Nmax,E,Emax);
+
+    struct node *node1 = malloc(sizeof(struct node));
+    struct node *node2 = malloc(sizeof(struct node));
+    struct node *temp = NULL;
+
+
+    /* Read the labels (and edge weight, if necessary) */
+    if (weighted == 0) {
+      noReadItems = sscanf(line, "%s %s", node1->label, node2->label);
+      weights[E] = 1;
+	  } else {
+      noReadItems = sscanf(line,"%s %s %lf", node1->label, node2->label, &weights[E]);
+    }
+
+    if (bipartite>1){
+      temp = node1;
+      node1 = node2;
+      node2 = temp;
+    }
+
+    /* Check input sanity */
+    if(noReadItems != noReadItemsExpected){
+      printf ("Failed to read input: not enough fields in line %s (%d!=%d). \n",
+              line, noReadItems,noReadItemsExpected);
+      return 1;
+    }
+
+    // Check if the first label was encountered already.
+    nodeptr = (struct node*) tsearch((void*) node1, (void **)&nodes_root, node_compare);
+    struct node *answer = NULL;
+    answer = *(struct node **)nodeptr;
+
+    // If the node was not present in the tree, save a new label,
+    // Otherwise free the read node to avoid memory leak.
+    if (answer==node1){
+      answer->id = N;
+      labels[N] = malloc(sizeof(answer->label));
+      strcpy(labels[N],answer->label);
+      N++;
+    }else{
+      free(node1);
+    }
+    nodes_in[E] = answer->id;
+
+    // Likewise for the second label
+    struct node *answer2 = NULL;
+    if(!bipartite){
+      nodeptr = (struct node*) tsearch((void*) node2, (void **)&nodes_root, node_compare);
+      answer2 = *(struct node **)nodeptr;
+      if (answer2==node2){
+        answer2->id = N;
+        labels[N] = malloc(sizeof(answer2->label));
+        strcpy(labels[N],answer2->label);
+        N++;
+      }else{
+        free(node2);
+      }
+    }
+    // For bipartite networks,
+    // - Each colums is a different namespace.
+    // - the node label of the second column are not saved.
+    else{
+      nodeptr = (struct node*) tsearch((void*) node2, (void **)&nodes_root_bipartite, node_compare);
+      answer2 = *(struct node **)nodeptr;
+      if (answer2==node2){
+        answer2->id = N_other_side;
+        N_other_side++;
+      } else{
+        free(node2);
+      }
+    }
+    nodes_out[E] = answer2->id;
+    E++;
+    // Increase the allocated memory if needed.
+    if(N+2>=Nmax){
+      printf("%d > %d, doubling memory for nodes\n",N,Nmax );
+      Nmax *= 2;
+      labels = realloc(labels, (Nmax+1)*sizeof(char*));
+    }
+    if(E+1>=Emax){
+      printf("%d > %d, doubling memory for edges",E,Emax );
+      Emax *= 2;
+      nodes_in = realloc(nodes_in, (Emax+1)*sizeof(unsigned int));
+      nodes_out = realloc(nodes_out, (Emax+1)*sizeof(unsigned int));
+      weights = realloc(weights, (Emax+1)*sizeof(double));
+    }
   }
-  else {  /* Node is a hub */
-	if (P < 0.300)
-	  dest_group = 4;
-	else if (P < 0.750)
-	  dest_group = 5;
-	else
-	  dest_group = 6;
-  }
-  return dest_group;
+
+  // Free unused memory
+  nodes_in = realloc(nodes_in, (E+1)*sizeof(unsigned int));
+  nodes_out = realloc(nodes_out, (E+1)*sizeof(unsigned int));
+  weights = realloc(weights, (E+1)*sizeof(double));
+  labels = realloc(labels, (N+1)*sizeof(char*));
+  free_node_tree(nodes_root);
+  free_node_tree(nodes_root_bipartite);
+
+  // Assign
+  *nodes_in_p = nodes_in;
+  *nodes_out_p = nodes_out;
+  *labels_p = labels;
+  *weights_p = weights;
+  *E_p = E;
+  *N_p = N;
+  free(line);
+
+  if (!bipartite)
+    printf("Read %d nodes and %d edges \n ", *N_p,*E_p);
+  else
+    printf("Read %d/%d nodes and %d edges \n ", *N_p,N_other_side,*E_p);
+  //
+  // for (i=0;i<*E_p;i++){
+  //   printf("%d\t [%d] %s \t [%d] %s \t %f\n",i, nodes_in[i],labels[nodes_in[i]],nodes_out[i],labels[nodes_out[i]],weights[i]);
+  // }
+  // for (i=0;i<*N_p+1;i++){
+  //   printf("%d %s\n",i, labels[i]);
+  // }
+  return 0;
 }
+
 
 
 void
@@ -62,9 +235,9 @@ ClusteringOutput(FILE *outf,
 				 Partition *part,
 				 char **labels){
   unsigned int mod;
-  Node *node;
-  for(mod=0;mod<part->M;mod++){
-	for(node = part->modules[mod]->first;node!=NULL; node=node->next){
+  Node *node = NULL;
+  for(mod=0; mod<part->M; mod++){
+	for(node = part->modules[mod]->first; node!=NULL; node=node->next){
 	  fprintf(outf,"%s\t",labels[node->id]);
 	}
 	fprintf(outf,"\n");
@@ -120,195 +293,4 @@ AssignNodesToModulesFromFile(FILE *inF,
   CompressPartition(part);
   hdestroy();
   return nnode;
-}
-
-/**
-Normalize edges weight and node strength and store them in the
-Partition and AdjaArray structures.
-
-This function assume that the edge list is UNDIRECTED and WITHOUT
-duplicates.
-
-If W_ij is the adjacency matrix and W=\sum_i\sum_j W_ij the sum of its
-elements. The normalized strength is:
-- For an edge:  A_ij = W_ij / W
-- For a node: k_i = sum_i (W_ij) / W
-
-@param nd_in,nd_out,weight Edge list, nodes id and weight of each edge.
-@param adj The adjacency array to fill.
-@param part The partition to fill.
-@param normalize If set to 0, the strength and weight are not normalized by W.
-**/
-int
-EdgeListToAdjaArray(int *nd_in, int *nd_out, double *weight,
-					AdjaArray *adj, Partition *part, int normalize){
-
-  unsigned int N = adj->N, E = adj->E;
-  double weightsum = 0;
-  unsigned int i;
-
-  int *degree = (int*) calloc(N,sizeof(int));
-  if (degree == NULL){
-	   return 1;
-  }
-
-  // Compute degrees and the sum of edges weight.
-  for (i=0;i<E;i++){
-  	weightsum += weight[i];
-
-  	part->nodes[nd_in[i]]->strength +=  weight[i];
-  	part->nodes[nd_out[i]]->strength +=  weight[i];
-
-  	degree[nd_in[i]] += 1;
-  	degree[nd_out[i]] += 1;
-  }
-
-  // If the weights and strength are not to be normalized, set the
-  // normalization constant to 1 (neutral element for division).
-  if (!normalize)
-	 weightsum = 1;
-
-  // Set the start of the index of the neighbors and store the
-  // normalized strength.
-  int pos = 0;
-  for (i=0;i<N;i++){
-  	adj->idx[i] = pos;
-  	pos += degree[i];
-  	part->nodes[i]->strength /=  weightsum;
-  }
-
-  // Fill the edges properties (target and normalized weights).  We
-  // re-use the degree array to keep track of the position to fill for
-  // each node. The first edge encountered for a node will be stored
-  // in the first position for this node:
-  // (idx[nodeid+1]-degree[node]=idx[nodeid]). Successives edges will
-  // be stored in successive positions (because degree[node] is
-  // decremented each time).
-  int idx_in, idx_out;
-  for (i=0;i<E;i++){
-  	// Get the positions to fill.
-  	idx_in  = adj->idx[nd_in[i]+1]-degree[nd_in[i]];
-  	idx_out = adj->idx[nd_out[i]+1]-degree[nd_out[i]];
-
-  	// Fill them with target and strength.
-  	adj->neighbors[idx_in] = nd_out[i];
-    adj->strength[idx_in] = weight[i]/weightsum;
-  	adj->neighbors[idx_out] = nd_in[i];
-  	adj->strength[idx_out] = weight[i]/weightsum;
-
-  	// Update the counter to fill the next position correctly.
-  	degree[nd_in[i]]--;
-  	degree[nd_out[i]]--;
-
-  }
-  return 0;
-}
-
-
-/**
-Compare Edges structures (used with Qsort)
-
-This function compare two edges structures starting by the target node
-(if they are equal it compare the source node).
-
-It returns 0 if the edges are identiqual or -1 (resp 1) if the first edge has
-smaller (resp. bigger) node ids.
-**/
-static int
-compare_edges(const void *p1, const void *p2)
-{
-        unsigned int y1 = ((Edge*)p1)->node2;
-        unsigned int y2 = ((Edge*)p2)->node2;
-        if (y1 < y2)
-          return -1;
-        else if (y1>y2)
-          return 1;
-        y1 = ((Edge*)p1)->node1;
-        y2 = ((Edge*)p2)->node1;
-        if (y1 < y2)
-          return -1;
-        else if (y1>y2)
-          return 1;
-        return 0;
-}
-
-
-
-/**
-Bipartite projection according to the second column
-**/
-unsigned int
-project_bipart(unsigned int *nd_in, unsigned int *nd_out, double *weights, int E,
-               unsigned int **proj1, unsigned int **proj2, double **projW){
-
-  // Sort the edges by the nodes to project with.
-  int i, count = E;
-  Edge *ed = malloc(count*sizeof(Edge));
-  for (i = 0; i < count; i++) {
-    ed[i].node1 = nd_in[i];
-    ed[i].node2 = nd_out[i];
-    ed[i].strength = weights[i];
-  }
-  qsort(ed, count, sizeof(Edge), compare_edges);
-
-  // Count the degree of each node and assign a first data structure for the
-  // projected network.
-  unsigned int degree = 0, Emax = 0;
-  for (i = 0; i <= count-1; i++) {
-    degree++;
-    if (ed[i].node2 !=ed[i+1].node2 || i == count-1) {
-      Emax += (degree)*(degree-1) / 2;
-      degree = 0;
-    }
-  }
-  Edge *projected = malloc(Emax*sizeof(Edge));
-
-  // Generate all projected edges
-  degree = 0;
-  unsigned int j = 0, x, y, x0 = 0;
-  for (i = 0; i <= count-1; i++) {
-    degree++;
-    if (ed[i].node2 !=ed[i+1].node2 || i == count-1) {
-      for (x = x0; x < i+1; x++){
-        for (y = x0; y < x; y++) {
-          if (ed[x].node1 < ed[y].node1){
-            projected[j].node1 = ed[x].node1;
-            projected[j].node2 = ed[y].node1;
-          }else{
-            projected[j].node1 = ed[x].node1;
-            projected[j].node2 = ed[y].node1;
-          }
-          projected[j].strength = ed[x].strength * ed[y].strength;
-          j++;
-        }
-      }
-      degree = 0;
-      x0 = i+1;
-  }
-  }
-  free(ed);
-  // Count the number of unique edges
-  qsort(projected, j, sizeof(Edge), compare_edges);
-  E = j;
-  for (i = 0; i < j-1; i++) {
-    if (compare_edges(&projected[i],&projected[i+1]) == 0)
-      E --;
-  }
-
-  // Fill the output array
-  *proj1 = malloc(E*sizeof(unsigned int));
-  *proj2 = malloc(E*sizeof(unsigned int));
-  *projW = calloc(E,sizeof(double));
-
-  unsigned int k = 0;
-  for (i = 0; i < j; i++) {
-    (*proj1)[k] = projected[i].node1;
-    (*proj2)[k] = projected[i].node2;
-    (*projW)[k] += projected[i].strength;
-    if (compare_edges(&projected[i],&projected[i+1]) != 0)
-      k++;
-    }
-  free(projected);
-
-  return E;
 }
