--- src/louvain.h
@@ -1,3 +1,0 @@
-#include "bipartite.h"
-
-struct group * LOUVCommunityIdentBipart(struct binet *binet, double epsilon, int weighted);
