+++ src/modules.h
@@ -174,6 +174,5 @@ struct group *CatalogRoleIdent(struct node_gra *net,
 struct group *CatalogRoleIdentStrength(struct node_gra *net,
 			       struct group *comm);
 
-int GetRole(double P, double z);
 
 #endif /* !RGRAPH_MODULES_H */
