+++ src/partition.h
@@ -1,9 +1,6 @@
-#ifndef __PARTITION_H__
-#define __PARTITION_H__
+#ifndef PARTITION_H__
+#define PARTITION_H__
 #include <gsl/gsl_rng.h>
-#include "graph.h"
-#include "modules.h"
-#include "bipartite.h"
 
 /**
 Adjacency Array.
@@ -70,49 +67,10 @@ typedef struct Partition {
 Partition * CreatePartition(unsigned int N, unsigned int M);
 Partition * CopyPartitionStruct(Partition *part);
 void FreePartition(Partition *part);
+
 void AssignNodesToModules(Partition *part, 	gsl_rng *gen);
-
-// Node movements
-void ChangeModule(unsigned int nodeId,
-				  unsigned int newModuleId,
-				  Partition *part);
-void MergeModules(unsigned int id1, unsigned int id2,
-				  Partition *part);
-unsigned int SplitModuleByComponent(unsigned int targetModuleId,
-									unsigned int emptyModuleId,
-									Partition *part, AdjaArray *adj,
-									gsl_rng *gen);
-// Modularity optimisation
-double dEChangeModule(unsigned int nodeid,
-					  unsigned int newModuleid,
-					  Partition *part, AdjaArray *adj);
-double dEMergeModules(unsigned int moduleId1,
-					  unsigned int moduleId2,
-					  Partition *part, AdjaArray *adj);
-
-
-/**
-A stack of integers, used to keep tracks of nodes to visit
-when traversing th graph.
-
-- The function AddToStack allows to add an item to the top of the
-  stack.
-- The function PopFromStack will return the item at the top of the
-  stack and remove it, or -1 if the stack is empty.
-**/
-typedef struct Stack {
-  unsigned int maxsize;
-  int top;
-  int *items;
-} Stack;
-
-Stack * CreateStack(unsigned int N);
-void FreeStack(Stack *st);
-void AddToStack(unsigned int value, Stack * st);
-int PopFromStack(Stack *st);
-
 double PartitionModularity(Partition *part, AdjaArray *adj, int diagonal_term);
 void PartitionRolesMetrics(Partition *part, AdjaArray *adj,  double connectivity[], double participartion[]);
 void CompressPartition(Partition *part);
-
+int GetRole(double P, double z);
 #endif
